filling na with zero for cases

+++ etl/steps/data/grapher/who/2023-06-30/guinea_worm.py
@@ -2,8 +2,7 @@
 
 from typing import cast
 
-import pandas as pd
-from owid.catalog import Dataset, Table
+from owid.catalog import Dataset
 
 from etl.helpers import PathFinder, create_dataset, grapher_checks
 
@@ -21,7 +20,7 @@ def run(dest_dir: str) -> None:
     # Read table from garden dataset.
     tb = ds_garden["guinea_worm"]
     tb["year_certified"] = tb["year_certified"].replace({"Pre-certification": 3000, "Endemic": 4000})
-    tb["year_certified"] = tb["year_certified"].astype("Int64")
+    tb["year_certified"] = tb["year_certified"].astype("float64")
     #
     # Process data.
     #
